decimal quantity

+++ food_db/food_db_app/models.py
@@ -63,7 +63,7 @@ class Ingredient(models.Model):
         null=True,
         blank=True,
     )
-    quantity = models.PositiveSmallIntegerField(null=True, blank=True)
+    quantity = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)
     ingredient_category = models.CharField(max_length=255, blank=True)
     notes = models.TextField(blank=True)
     _date_created = models.DateTimeField(auto_now_add=True)
